backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.35
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
bcrypt==4.0.1
python-multipart==0.0.22
rembg[cpu]
pillow
httpx
google-cloud-aiplatform
python-dotenv
google-cloud-aiplatform google-genai
